# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Backend API",
        "version": "1.0.0"
    },
    "basePath": "/api/",
    "paths": {
...
        "/assets/{id}": {
            "get": {
                "tags": [
                    "Asset"
                ],
                "summary": "Get asset",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target asset.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Asset"
                ],
                "summary": "Get assets",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target asset.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Asset",
                        "required": true,
                        "schema": {
                            "properties": {
                                "asset_type_id": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "currency_id": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "bank_id": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "name": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "address": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "key": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "default": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "notes": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Asset"
                ],
                "summary": "Get asset",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target asset.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/asset_types": {
            "get": {
                "tags": [
                    "Asset_type"
                ],
                "summary": "Get asset_types",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Asset_type"
                ],
                "summary": "create asset_type",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Asset",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Store a newly created resource in storage.",
                                    "type": "string"
                                },
                                "crypto": {
                                    "description": "Store a newly created resource in storage.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/asset_types/{id}": {
            "get": {
                "tags": [
                    "Asset_type"
                ],
                "summary": "Get asset_type",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target asset_type.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Asset_type"
                ],
                "summary": "Get asset_types",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target asset.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Asset_type",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "crypto": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Asset_type"
                ],
                "summary": "delete asset",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target asset_type.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/banks/{id}": {
            "get": {
                "tags": [
                    "banks"
                ],
                "summary": "Get bank",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target bank.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "banks"
                ],
                "summary": "Get banks",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target bank.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "bank",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "banks"
                ],
                "summary": "delete bank",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target banks.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/currencies/{id}": {
            "get": {
                "tags": [
                    "currencies"
                ],
                "summary": "Get currency",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target currency.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "currencies"
                ],
                "summary": "delete currency",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target currency.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/deals/{id}": {
            "get": {
                "tags": [
                    "deals"
                ],
                "summary": "Get deal",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target deal.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "deals"
                ],
                "summary": "update deal",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target deal.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "deal",
                        "required": true,
                        "schema": {
                            "property": "source_asset_id",
                            "description": "Update the specified resource in storage.",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "deals"
                ],
                "summary": "delete deal",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target deal.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/deal_stages/{id}": {
            "get": {
                "tags": [
                    "deal_stages"
                ],
                "summary": "Get deal_stages",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target deal_stages.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "deal_stages"
                ],
                "summary": "update deal_stages",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target deal_stages.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "deal_stages",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "deal_stages"
                ],
                "summary": "delete deal_stages",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Target deal_stages.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "JWT-token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}